+++ src/data/views.py
@@ -7,11 +7,10 @@ from drf_spectacular.utils import extend_schema
 from data.models import GeoData
 from data.serializers import (
     GeoDataSerializer,
-)
-from data.websocket.serializers import (
     AddressLookupSerializer,
-    AddressLookupResponseSerializer,
 )
+from data.websocket.serializers import AddressLookupResponseSerializer
+
 from data.tasks import process_geodata
